Move apt-specific classes to apt package
